backend/requirements.txt
Django==3.2.6
djangorestframework==3.11.1
djangorestframework_simplejwt==4.0.0
PyJWT==1.7.0
psycopg2-binary==2.8.5
django-cors-headers==3.7.0
django-model-utils==4.0.0
gunicorn==20.1.0
stripe==2.50.0
dj-stripe==2.4.4
social-auth-app-django==4.0.0
django-templated-mail==1.1.1
python-decouple==3.4
django-filter==2.3.0




